+++ android/app/build.gradle
@@ -40,9 +40,9 @@ android {
     defaultConfig {
         applicationId "de.delusionsoftware.congress_fahrplan"
         minSdkVersion flutter.minSdkVersion
-        targetSdkVersion 31
-        versionCode 2023
-        versionName '1.3.0'
+        targetSdkVersion 33
+        versionCode 0234
+        versionName '1.3.2'
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
