+++ android/app/src/debug/AndroidManifest.xml
@@ -8,5 +8,4 @@
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE"/>
     <uses-permission android:name="android.permission.READ_CALENDAR" />
     <uses-permission android:name="android.permission.WRITE_CALENDAR" />
-
 </manifest>
