+++ lib/provider/favorite_provider.dart
@@ -9,7 +9,7 @@ import 'dart:collection';
 
 import 'package:congress_fahrplan/model/fahrplan.dart';
 import 'package:congress_fahrplan/utilities/fahrplan_fetcher.dart';
-import 'package:congress_fahrplan/widgets/talk.dart';
+import 'package:congress_fahrplan/model/talk.dart';
 import 'package:flutter/material.dart';
 import 'package:package_info_plus/package_info_plus.dart';
