+++ lib/widgets/sync_calendar.dart
@@ -10,7 +10,7 @@ import 'dart:core';
 
 import 'package:congress_fahrplan/model/day.dart';
 import 'package:congress_fahrplan/provider/favorite_provider.dart';
-import 'package:congress_fahrplan/widgets/talk.dart';
+import 'package:congress_fahrplan/model/talk.dart';
 import 'package:device_calendar/device_calendar.dart';
 import 'package:flutter/material.dart';
 import 'package:timezone/data/latest.dart' as tz;
