+++ pubspec.yaml
@@ -16,7 +16,7 @@ description: Congress Fahrplan written in flutter
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.3.0
+version: 1.3.2
 
 environment:
   sdk: ">=2.14.0 <3.0.0"
